Renamed packages from "com.github.cloudyrock.springboot.v2_2" to "com.github.cloudyrock.springboot.v2_4" in module "mongock-springboot-v2.4".
