+++ integration-tests/src/main/java/com/github/cloudyrock/mongock/integrationtests/spring5/springdata3/events/MongockStartedEventListener.java
@@ -1,6 +1,6 @@
 package com.github.cloudyrock.mongock.integrationtests.spring5.springdata3.events;
 
-import com.github.cloudyrock.springboot.v2_2.events.SpringMigrationStartedEvent;
+import com.github.cloudyrock.springboot.v2_4.events.SpringMigrationStartedEvent;
 import org.springframework.context.ApplicationListener;
 import org.springframework.stereotype.Component;
