+++ integration-tests/src/test/java/com/github/cloudyrock/mongock/integrationtests/spring5/springdata3/RunnerLegacyMigrationITest.java
@@ -4,7 +4,7 @@ import com.github.cloudyrock.mongock.config.LegacyMigration;
 import com.github.cloudyrock.mongock.integrationtests.spring5.springdata3.changelogs.empty.EmptyChangeLog;
 import com.github.cloudyrock.mongock.integrationtests.spring5.springdata3.util.Constants;
 import com.github.cloudyrock.mongock.integrationtests.spring5.springdata3.util.LegacyMigrationUtils;
-import com.github.cloudyrock.springboot.v2_2.MongockSpringbootV2_4;
+import com.github.cloudyrock.springboot.v2_4.MongockSpringbootV2_4;
 import com.mongodb.client.MongoCollection;
 import org.bson.Document;
 import org.junit.jupiter.params.ParameterizedTest;
